Remove CSRF protection in register form
This protection ensures (using a generated token) that no external forms
are sent (which is what we want in this case!).

Usually this protection is recommended for forms sent by already
registered users (to avoid identity impersonation) or, less commonly,
login forms (to avoid trusted domain phishing attacks). We're still
protected against both, because we're only disabling CSRF in register
form.